v25

+++ IkinciElAracIhaleSistemi.UI/Controllers/AracController.cs
@@ -24,6 +24,7 @@ namespace IkinciElAracIhaleSistemi.UI.Controllers
 		[HttpGet]
 		public ActionResult AracEkle()
 		{
+
 			AracOzellikleriCache();
 			return View();
 		}
@@ -94,12 +95,12 @@ namespace IkinciElAracIhaleSistemi.UI.Controllers
 		/// </summary>
 		public void AracOzellikleriCache()
 		{
-			ViewBag.VitesTipleri = CacheHelper.GetOrSet("VitesTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.VitesTipi), DateTimeOffset.Now.AddMinutes(30));
-			ViewBag.GovdeTipleri = CacheHelper.GetOrSet("GovdeTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.GovdeTipi), DateTimeOffset.Now.AddMinutes(30));
-			ViewBag.YakitTipleri = CacheHelper.GetOrSet("YakitTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.YakitTipi), DateTimeOffset.Now.AddMinutes(30));
-			ViewBag.VersiyonTipleri = CacheHelper.GetOrSet("VersiyonTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.Versiyon), DateTimeOffset.Now.AddMinutes(30));
-			ViewBag.Renkler = CacheHelper.GetOrSet("Renkler", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.Renk), DateTimeOffset.Now.AddMinutes(30));
-			ViewBag.Donanimlar = CacheHelper.GetOrSet("Donanim", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.OpsiyonelDonanim), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.VitesTipleri = CacheHelper.GetOrSet("VitesTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.VitesTipi), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.GovdeTipleri = CacheHelper.GetOrSet("GovdeTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.GovdeTipi), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.YakitTipleri = CacheHelper.GetOrSet("YakitTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.YakitTipi), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.VersiyonTipleri = CacheHelper.GetOrSet("VersiyonTipleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.Versiyon), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.Renkler = CacheHelper.GetOrSet("Renkler", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.Renk), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.Donanimlar = CacheHelper.GetOrSet("Donanim", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.OpsiyonelDonanim), DateTimeOffset.Now.AddMinutes(30));
 			
 			ViewBag.Firmalar = CacheHelper.GetOrSet("Firmalar", () => new FirmaDAL().FirmaListesineDonustur(), DateTimeOffset.Now.AddMinutes(30));
 			ViewBag.BireyselUyeler = CacheHelper.GetOrSet("BireyselUyeler", () => new BireyselUyeDAL().BireyselUyeleriListeyeDonustur(UyeTurleri.Bireysel), DateTimeOffset.Now.AddMinutes(30));
@@ -109,7 +110,7 @@ namespace IkinciElAracIhaleSistemi.UI.Controllers
 
 		public void AracListelemeOzellikleriCache()
 		{
-			ViewBag.AracTurleri = CacheHelper.GetOrSet("AracTurleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur(AracOzellikleri.AracTuru), DateTimeOffset.Now.AddMinutes(30));
+			ViewBag.AracTurleri = CacheHelper.GetOrSet("AracTurleri", () => new AracOzellikDAL().AracOzellikleriniListeyeDonustur((int)AracOzellikleri.AracTuru), DateTimeOffset.Now.AddMinutes(30));
 			ViewBag.Modeller = CacheHelper.GetOrSet("Modeller", () => new ModelDAL().ModelListesineDonustur(), DateTimeOffset.Now.AddMinutes(30));
 			ViewBag.Markalar = CacheHelper.GetOrSet("Markalar", () => new MarkaDAL().MarkaListesineDonustur(), DateTimeOffset.Now.AddMinutes(30));
 			ViewBag.Statuler = CacheHelper.GetOrSet("Statuler", () => new StatuDAL().StatuListesineDonustur(), DateTimeOffset.Now.AddMinutes(30));
